removed BackgroundWorker, cleaned up
